mp3 van sounds en upgrade turrets

+++ Assets/Scripts/Enemy.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using UnityEngine;
 
 public class Enemy : MonoBehaviour {
-    public int health;
+    public float health;
     public bool isDead = false;
     public int moneyAmount;
     public int damage;
@@ -18,7 +18,7 @@ public class Enemy : MonoBehaviour {
         Destroy(gameObject);
     }
 
-    public void LoseLife(int life)
+    public void LoseLife(float life)
     {
         health -= life;
